requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
huggingface_hub>=0.22.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
flask>=3.0.0
